LayoutTests/editing/deleting/5390681-expected.txt

This tests for a bug where expansion for smart delete would not consider editable boundaries.  Only 'foo' should be deleted.  You should see ' bar'.  There is a bug: while the non-editable space isn't deleted, deletion inserts a placeholder when it shouldn't.
| <br>
| <span>
|   contenteditable="false"
|   " bar"
